--- resource_logger/Dockerfile/resource_list.json
@@ -1,7 +1,0 @@
-[
-    {
-        "Group":   "samplecontroller.k8s.io",
-        "Version": "v1alpha1",
-        "Kind":    "Foo"
-    }
-]
