--- resource_logger/Dockerfile/resource_list_testing.json
@@ -1,15 +1,0 @@
-[
-    {
-        "Version": "v1",
-        "Kind":    "Service"
-    },
-    {
-        "Version": "v1",
-        "Kind":    "Pod"
-    },
-    {
-        "Group":   "samplecontroller.k8s.io",
-        "Version": "v1alpha1",
-        "Kind":    "Foo"
-    }
-]
