+++ watcher/Dockerfile/resource_list.json
@@ -1,0 +1,7 @@
+[
+    {
+        "Group":   "samplecontroller.k8s.io",
+        "Version": "v1alpha1",
+        "Kind":    "Foo"
+    }
+]
